Renamed collections.py to collection.py.
