+++ src/imagedata/collection.py
