+++ tests/test_formats_itk.py
@@ -14,7 +14,7 @@ import argparse
 import src.imagedata.cmdline as cmdline
 import src.imagedata.formats as formats
 from src.imagedata.series import Series
-from src.imagedata.collections import Cohort
+from src.imagedata.collection import Cohort
 from .compare_headers import compare_headers
