+++ tests/test_formats_mat.py
@@ -9,7 +9,7 @@ import argparse
 import src.imagedata.cmdline as cmdline
 import src.imagedata.formats as formats
 from src.imagedata.series import Series
-from src.imagedata.collections import Cohort
+from src.imagedata.collection import Cohort
 
 
 def list_files(startpath):
